+++ Libplanet.Explorer/Queries/TransactionQuery.cs
@@ -36,11 +36,6 @@ namespace Libplanet.Explorer.Queries
                         Name = "signer",
                         DefaultValue = null,
                     },
-                    new QueryArgument<AddressType>
-                    {
-                        Name = "involvedAddress",
-                        DefaultValue = null,
-                    },
                     new QueryArgument<BooleanGraphType>
                     {
                         Name = "desc",
@@ -56,12 +51,11 @@ namespace Libplanet.Explorer.Queries
                 resolve: context =>
                 {
                     var signer = context.GetArgument<Address?>("signer");
-                    var involved = context.GetArgument<Address?>("involvedAddress");
                     bool desc = context.GetArgument<bool>("desc");
                     long offset = context.GetArgument<long>("offset");
                     int? limit = context.GetArgument<int?>("limit");
 
-                    return ExplorerQuery.ListTransactions(signer, involved, desc, offset, limit);
+                    return ExplorerQuery.ListTransactions(signer, desc, offset, limit);
                 }
             );
 
@@ -73,11 +67,6 @@ namespace Libplanet.Explorer.Queries
                         Name = "signer",
                         DefaultValue = null,
                     },
-                    new QueryArgument<AddressType>
-                    {
-                        Name = "involvedAddress",
-                        DefaultValue = null,
-                    },
                     new QueryArgument<BooleanGraphType>
                     {
                         Name = "desc",
@@ -93,14 +82,12 @@ namespace Libplanet.Explorer.Queries
                 resolve: context =>
                 {
                     var signer = context.GetArgument<Address?>("signer");
-                    var involved = context.GetArgument<Address?>("involvedAddress");
                     bool desc = context.GetArgument<bool>("desc");
                     int offset = context.GetArgument<int>("offset");
                     int? limit = context.GetArgument<int?>("limit", null);
 
                     return ExplorerQuery.ListStagedTransactions(
                         signer,
-                        involved,
                         desc,
                         offset,
                         limit